// bake/schema/openapi-types.ts
interface Map<T> {
  [key: string]: T;
}
// ...
export interface ExternalDocumentationObject {
  description?: string;
  url: string;
}
// ...
export type JSONSchema = JSONSchema4 | JSONSchema6 | JSONSchema7;
export type SchemaObject =
  & Omit<
    JSONSchema,
    | "type"
    | "allOf"
    | "oneOf"
    | "anyOf"
    | "not"
    | "items"
    | "properties"
    | "additionalProperties"
  >
  & {
    // OpenAPI specific overrides
    type?: "string" | "number" | "integer" | "boolean" | "object" | "array";
    allOf?: SchemaObject[];
    oneOf?: SchemaObject[];
    anyOf?: SchemaObject[];
    not?: SchemaObject;
    items?: SchemaObject;
    properties?: Map<SchemaObject>;
    additionalProperties?: Map<SchemaObject>;

    // OpenAPI additions
    nullable?: boolean;
    discriminator?: DiscriminatorObject;
    readOnly?: boolean;
    writeOnly?: boolean;
    xml?: XMLObject;
    externalDocs?: ExternalDocumentationObject;
    example?: any;
    deprecated?: boolean;

    // For ref tracking
    "#ref"?: string;
  };
// ...
export interface DiscriminatorObject {
  propertyName: string;
  mapping?: Map<string>;
}

export interface XMLObject {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;
}